bug fix for empty folder count in summary
Fixed a bug in the user summary wherein empty folders were counted as those with no files (but possibly subfolders, i.e., not really empty). This was causing an inflated (inaccurate) count of empty folders. Now redefined as folders with no files or subfolders. Thanks to an anonymous CHI`19 reviewer for helping to identify this bug.